refactor: More changes
